Move test cases out of serialize and `pub` the module
